+++ src/utils/libs/validate.js
@@ -53,7 +53,6 @@ function isEmpty(input) {
         return !input.replace(/\s/g, '').length === true;
     }
 }
-
 function isNumber(input) {
     const reg = /^[\d|\.|,]+$/;
     return reg.test(input);